Move vocab functions to vocabulary_manager
